Update list of my online profiles.
The "Video channels" list is now an unordered list, periods have been
removed from the end of the list items, and my Vimeo channel has been
added to the list.

Also added when the page was last updated since I don't really want that
to be on each post at the top if they haven't changed.